Here is a code online price comparision for online shopping sites
There are 2 types of
comparision imlemented

1.site comparision
comparision  where we can enter  product and compares price in different
shopping sites.

2. product comparision
where we choose one shopping site and compares 2 products(brands) in
that site.

It is implemented for   3 online shopping sites namely amazon,ebay and
junglee.The whole system is implemented in html and php in front end and
mysql wamp server in back end.